backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
langchain==0.1.16
langgraph==0.0.40
anthropic==0.25.0
groq==0.8.0
supabase==2.4.0
python-dotenv==1.0.1
sse-starlette==1.8.0
pydantic==2.6.0
